Moved scraper.py into a project directory and linted
